requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.8
Flask-Caching==1.9.0
Flask-Migrate==2.5.3
mysqlclient==1.4.6
dateparser==1.0.0
